Updated Public & Secret backend to use numba/cuda

+++ utils/const.py
@@ -1,5 +1,5 @@
 import numpy
 
-INT = numpy.int64
+INT = numpy.int32
 
 MAX_CHR = 1114111
